scripts/requirements.txt
supabase==2.9.1
python-dotenv==1.0.0
httpx==0.27.0
